requirements.txt
google-adk==1.4.1
google-generativeai==0.8.5
python-dotenv==1.1.0
fastapi
uvicorn[standard]